Automatically backed up by Learn

+++ lib/create.sql
@@ -15,7 +15,7 @@ CREATE TABLE projects (
   
   CREATE TABLE pledges ( 
     id INTEGER PRIMARY KEY,
-    amount REAL,
+    amount INTEGER,
     user_id INTEGER,
     project_id INTEGER
   );
